+++ src/onegov.search/HISTORY.rst
@@ -4,7 +4,10 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
-- Add a switch to disable elasticsearch integration.
+- Adds the ability to define multiple suggestion inputs per document.
+  [href]
+
+- Adds a switch to disable elasticsearch integration.
   [href]
 
 0.0.5 (2015-09-25)